Moved AODVTest out of test.h/cc so it can be used in other places. REM tests passing
