+++ aodv/Makefile
@@ -21,10 +21,10 @@ LIBFLAGS = -cvq
 #
 TARGET = test
 LIB_NAME = libaodv.a
-OBJ_LIB_FILES = aodv.o aodv_rreq.o aodv_rrep.o aodv_rerr.o aodv_routing_table.o ../adhoc/*.o
+OBJ_LIB_FILES = aodv.o aodv_rreq.o aodv_rrep.o aodv_rerr.o aodv_routing_table.o aodv_test.o ../adhoc/*.o
 
 OBJ_FILES = ${TARGET}.o ${OBJ_LIB_FILES}
-INC_FILES = ${TARGET}.h aodv.h aodv_rreq.h aodv_rrep.h aodv_rerr.h aodv_routing_table.h aodv_params.h defines.h network_monitor.h monitor_info.h
+INC_FILES = ${TARGET}.h aodv.h aodv_rreq.h aodv_rrep.h aodv_rerr.h aodv_routing_table.h aodv_test.h aodv_params.h defines.h network_monitor.h monitor_info.h
 
 SUBDIRS = ../adhoc
